OM-1 Update calls for renamed architectural folders
