+++ server/adapters/controllers/orders/getOrderById.ts
@@ -3,7 +3,7 @@
 // @access Public
 import asyncHandler from "express-async-handler";
 import { BY_ID } from "../../config/constants";
-import db from "../../../presentation/db/models";
+import db from "../../../infrastructure/db/models";
 import { getOrderById as useCase } from "../../../application/useCases";
 
 const { Order } = db;
